projects/time_of_day_clustering/project-requirements.txt
./dist/january_ml-0.0.1-py3-none-any.whl
dotenv
snowflake-snowpark-python
snowflake-ml-python
scikit-learn